(GEP-189) Allow multiple validation steps
The Validation and Publication action had an action specific URL
(as most actions do) which caused the reports to be overwritten
when more than one action of the same type was added to the same job.

This commit ensures that the URL and the action description of both
a validation and a publication step will be unique when more than
one validation step is added to a configuration.