(PF-828) Validate what versions of puppet a module will work with
This commit enables entry of two puppet versions (min and max) and
changes the logic so that all versions in the given range are
validated. It also adds an extra field where a json output file
can be entered. When present, that file will receive JSON output
for all validated versions.